todo-fastmcp/pyproject.toml
[project]
name = "todo-fastmcp"
version = "0.1.0"
description = "FastMCP-based MCP server providing in-memory todo CRUD tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.13.0.2",
    "python-dotenv>=1.0.0"
]
